ai/requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
httpx>=0.23.0


